Set SDK to Yosemite
Grabbed version on Fri Jan 16 15:49:05 2015 +0000 from
svn://svn.code.sf.net/p/chicken/code/trunk

Then modified all sdk versions to the latest XCode as of that date to
test.

Compiled, ran, and found that many bugs in 2.2b2 were fixed in master.

Note that I know very little about XCode so I'm sure that a release
could be done in a more platform-generic way for a release.

I would love to see a proper release done, but if you need a working vnc
on a Mac and you have current XCode, this could help you.